Move ApplicationComponent to Matey namespace to avoid naming conflicts
