+++ app/components/matey/active_users_component.rb
@@ -1,6 +1,6 @@
 require "ahoy_matey"
 
-class Matey::ActiveUsersComponent < ApplicationComponent
+class Matey::ActiveUsersComponent < Matey::ApplicationComponent
   def initialize(events:, time_window: 1.week)
     raise ArgumentError unless events.is_a?(ActiveRecord::Relation)
     raise ArgumentError unless time_window.is_a?(Integer)
