+++ app/components/matey/application_component.rb
@@ -1,7 +1,7 @@
 require "view_component"
 require "ahoy_matey"
 
-class ApplicationComponent < ViewComponent::Base
+class Matey::ApplicationComponent < ViewComponent::Base
   include ActiveModel::Validations
 
   def before_render
