+++ app/components/matey/bounce_rate_component.rb
@@ -1,6 +1,6 @@
 require "ahoy_matey"
 
-class Matey::BounceRateComponent < ApplicationComponent
+class Matey::BounceRateComponent < Matey::ApplicationComponent
   def initialize(events:, visits:, limit: 5)
     # Determine the total number of user sessions to the website
     @total_number_of_user_visits = events.pluck(:visit_id).uniq.count
