+++ app/components/matey/browser_os_breakdown_component.rb
@@ -1,4 +1,4 @@
-class Matey::BrowserOsBreakdownComponent < ApplicationComponent
+class Matey::BrowserOsBreakdownComponent < Matey::ApplicationComponent
   def initialize(visits:, time_window:)
     visits_in_time_window = visits.where(started_at: time_window.ago..)
     @visits_in_time_window = visits_in_time_window.count
