+++ app/components/matey/daily_active_users_component.rb
@@ -1,4 +1,4 @@
-class Matey::DailyActiveUsersComponent < ApplicationComponent
+class Matey::DailyActiveUsersComponent < Matey::ApplicationComponent
   def initialize(visits:, time_window:)
     @visits = visits
     @time_window = time_window
