+++ app/components/matey/new_users_component.rb
@@ -1,4 +1,4 @@
-class Matey::NewUsersComponent < ApplicationComponent
+class Matey::NewUsersComponent < Matey::ApplicationComponent
   def initialize(users:, time_window: 1.week)
     @current_period = users.where(created_at: time_window.ago..Time.current).count
     previous_period = users.where(created_at: (2 * time_window).ago..time_window.ago).count
