+++ app/components/matey/top_visited_pages_table_component.rb
@@ -1,4 +1,4 @@
-class Matey::TopVisitedPagesTableComponent < ApplicationComponent
+class Matey::TopVisitedPagesTableComponent < Matey::ApplicationComponent
   def initialize(events:, time_window: 1.week, limit: 10)
     # Group events by controller (:name) and action. Aggregate number of unique user actions
     @user_count_by_event = events.where(started_at: time_window.ago..).pluck(:landing_page).tally
